Refactor of CoursesForm
Now is not necessary to pass an array of courses to the CoursesForm in order to filter the courses and show the selected ones either for Students or Teacher. Since the CoursesForm is fetching all the courses and each course has the information about teacher and/or students we only need to pass the id (teacher or student) and the component will filter the courses and show the selected ones.