Refactor of CoursesForm
Going back with change e3cd9cb98afc1c9abdf90077e86286d64ee3e843 since it seems to be a more natural way to pass the markedCourses into the component in order to show them checked.
Now we to pass an array of courses id to the CoursesForm in order to filter the courses and show the selected ones either for Students or Teacher.