Add visual consistency between courses toggle status indicator and course status on courses view
Before this change on courses view when admin clicks on status toggle button it prompts a confirmation modal and if modal is canceled the toggle indicator showed the new toggled statues.
After this change the toggle will keep showing the current status of the course even after is clicked and it will change only after modal confirmation and page refresh.