Improve performance when add/remove students from a course
Before this change when admin added/removed students to a Course the logic was to remove the course form all student registered to that course and register the new ones. For instance: if course had 80 students and you want to remove only 1, it was removing the 80 students and adding 79, -> 159 transactions
After this change when admin adds/remove a student the service only adds the new ones and remove the needed students. For instance: if course had 80 students and you want to remove only 1, it only removes 1 student -> transaction