Improve performance when add/remove courses to/from a student
Before this change when admin added/removed courses to a Student the logic was to remove the student form all courses registered to that student and register the new ones. For instance: if student had registered in 10 courses and you want to remove only 1, it was removing the 10 students and adding 9, -> 19 transactions
After this change when admin adds/remove a course to a student the service only adds the new ones and remove the unselecteds. For instance: if student had registered in 10 courses and you want to remove only 1, it will remove only 1, -> 1 transactions