Improve performance when add/remove teachers to/from a course
Before this change when admin added/removed courses to a Teacher the logic was to remove the teacher from all courses assigned to that Teacher and register the new ones. For instance: if Teacher was registered in 10 courses and you want to remove only 1, it was removing the 10 Teacher and adding 9, -> 19 transactions
After this change when admin adds/remove a course to a Teacher the service only adds the new ones and remove the unselected. For instance: if Teacher was registered in 10 courses and you want to remove only 1, it will remove only 1, -> 1 transactions