Change unsubscribe() strategy
This will implement a more declarative way of unsubscribe from observables on the components. Instead of doing an unsubscribe() when the component is destroyed now we take the observables while the component is alive -> takeWhile(() => this.isAlive) and on destroy instead of doing all unsubscribes we only set isAlive to false.