Little Login implementation refactor
This will change the login implementation in order to authorize user (setToken) only after retrieve the full account info. Now login will be complete only after fetch the Account info, and after this the token will be set. This is to prevent some corner cases where the user could be authorized but app will not now if the user profile is administrator or not because this info is contained within the account info. Also AuthGuard will validate the user has a local token and if there is no account it will pull the info doing a loginBytoken